services/eval/pyproject.toml
[tool.poetry]
name = "eval"
version = "0.1.0"
description = "AI/OCR Evaluation Suite for Palateful"
authors = ["Leonid Belyi"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
pyyaml = "^6.0"
click = "^8.1"
rich = "^14.0"
jinja2 = "^3.1"
python-dotenv = "^1.1"
python-Levenshtein = "^0.27"
nltk = "^3.9"
sqlalchemy = "^2.0"
asyncpg = "^0.30"
openai = "^2.8"
httpx = "^0.27"
pillow = ">=11.3.0"
utils = {path = "../../libraries/utils", develop = true}

[tool.poetry.group.dev.dependencies]
ruff = "^0.11"
pytest = "^8.0"
pytest-cov = "^6.0"

[tool.poetry.scripts]
eval = "src.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
